requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
aiohttp==3.9.3
websockets==12.0
groq==0.4.2
python-multipart==0.0.9
jinja2==3.1.3
